fix(web): derive runtime-log subject keys from a storage-free leaf module
