+++ tsconfig.base.json
@@ -165,8 +165,8 @@
       "#hosted-web-testing": [
         "./apps/web/test/support/hosted-web-testkit.ts"
       ],
-      "@/src/lib/hosted-runtime-log/store": [
-        "./apps/web/src/lib/hosted-runtime-log/store.ts"
+      "@/src/lib/hosted-runtime-log/subject-key": [
+        "./apps/web/src/lib/hosted-runtime-log/subject-key.ts"
       ],
       "@murphai/messaging-ingress/linq-webhook": [
         "./packages/messaging-ingress/src/linq-webhook.ts"
